factored out the matches & notifs routes to match the refactor
