+++ app.js
@@ -97,6 +97,9 @@ app.use(methodOverride('_method'));
 app.use('/dashboard', require('./routes/ads.js'));
 app.use('/profile', require('./routes/profileInfo.js'));
 app.use('/user-profile', require('./routes/userProfile.js'));
+app.use('/matches', require('./routes/matches.js'));
+app.use('/notifications', require('./routes/notifications.js'));
+
 app.get('/create-profile',utils.checkAuthenticated,(req,res) => {
     mysql.pool.query("CALL GetInstrumentsLevels()", [], (error, rows) => {
         if(error) {
@@ -275,183 +278,6 @@ app.post('/register', utils.checkNotAuthenticated, async function (req, res) {
 app.delete('/logout', (req, res) => {
     req.logOut();
     res.redirect('/'); //on logout, redirect to landing page
-});
-
-
-
-app.get('/matches',utils.checkAuthenticated,(req, res, next) => {
-    // Get all matches in the matches table, then:
-    res.render('matches', {
-        user: req.user, notifs: req.session.notifs, unreadNotifs: req.session.unreadNotifs, profile: true });
-});
-
-app.get('/matches/pending',utils.checkAuthenticated,(req, res, next) => {
-   try {
-       let sql = 'SELECT * FROM Matches WHERE Accepted = 0 AND MatchedProfileID = ?';
-       mysql.pool.query(sql, [req.session.ProfileID], function (err, result) {
-           if(err) {
-               throw(err);
-           } else {
-               res.send(JSON.stringify(result));
-           }
-       });
-   } catch(err) {
-       res.redirect(utils.errorRedirect('/matches/pending', 'An unexpected error occurred retrieving your matches'));
-   }
-});
-
-
-//adds a new match with pending status and adds notification for user to accept/reject
-app.post('/matches/add', utils.checkAuthenticated, (req, res, next) => {
-    try {
-        mysql.pool.getConnection(function (err, conn) {
-            if (err) throw (err);
-
-            //add the match record
-            conn.query('INSERT INTO Matches (AdID, MatchedProfileID, Accepted, DateInviteSent) VALUES (?, ?, ?, NOW())',
-                [req.body.AdID, req.body.MatchedProfileID, false],
-                function (err, rows) {
-
-                    if (err) { //something went wrong so send an error
-                        conn.release();
-                        res.write({ success: 0, message: 'An error occurred when requesting to match: '.JSON.stringify(err) });
-                        res.end();
-                    } else {
-
-                        var matchKey = rows.insertId; //id of match we just created.
-
-                        //now need the notification record for the person who owns the profile
-
-                        //first go get the userID of the profile ID
-                        conn.query(`SELECT UserID FROM Profiles WHERE ProfileKey = ? LIMIT 1`, [req.body.MatchedProfileID],
-                            function (err, rows) {
-
-                                if (err) {
-                                    conn.release();
-                                    res.write({ success: 0, message: 'An error occurred when getting user connected to profile: '.JSON.stringify(err) });
-                                    res.end();
-                                } else if (rows.length === 0) {
-                                    conn.release();
-                                    res.write({ success: 0, message: 'Profile user does not exist. ' }); //no user found! No goood
-                                    res.end();
-                                }
-                                else {
-                                    var userID = rows[0]["UserID"];
-
-                                    //now go add the notification record.
-                                    conn.query(`INSERT INTO Notifications (UserID, MatchID, Msg, ReadMsg, CreateDate) VALUES (?, ?, ?, ?, NOW()) `,
-                                        [userID, matchKey, "New match request from <strong>" + req.user.FirstName + " " + req.user.LastName +"</strong>! Click to view your matches now.", false],
-                                        function (err, rows) {
-                                            conn.release();
-
-                                            if (err) {
-                                                res.write({ success: 1, message: 'An error occurred adding notification for match request: '.JSON.stringify(err) });
-                                                res.end();
-                                            } else res.send({ success: 1, message: 'Successfully requested to match.' });
-                                        });
-                                }
-                            });
-                    }
-                });
-
-        });
-    } catch (err) {
-        res.write(JSON.stringify(err));
-        res.end();
-    }
-});
-
-app.put('/notifications/markRead',utils.checkAuthenticated,(req, res, next) => {
-
-    try {
-        let sql = 'UPDATE Notifications SET ReadMsg=1 WHERE UserID=?';
-        mysql.pool.query(sql, [req.user.UserKey], function (err, result) {
-           if(err) {
-               throw(err);
-           } else if (result.affectedRows >= 1) {
-                res.send(true);
-           } else {
-               throw(new ReferenceError('No unread notifications found'));
-           }
-        });
-    } catch(err) {
-        res.redirect(utils.errorRedirect('/notifications/markRead', 'An unexpected error occurred marking your notifications read'));
-    }
-});
-
-
-app.put('/notifications/markRead/:id',utils.checkAuthenticated,(req, res, next) => {
-
-    try {
-        let sql = 'UPDATE Notifications SET ReadMsg=1 WHERE NotificationKey=?';
-        mysql.pool.query(sql, [req.params.id], function(err, result) {
-           if(err) {
-               throw(err);
-           } else if (result.affectedRows === 1) {
-               res.send(true);
-           } else {
-               throw(new ReferenceError('No such notification or notification already read'));
-           }
-        });
-    } catch(err) {
-        res.redirect(utils.errorRedirect('/notifications/markRead', 'An unexpected error occurred marking your notification read'));
-    }
-});
-
-
-app.put('/notifications/markUnread/:id', utils.checkAuthenticated, (req, res, next) => {
-
-    try {
-        let sql = 'UPDATE Notifications SET ReadMsg=0 WHERE NotificationKey=?';
-        mysql.pool.query(sql, [req.params.id], function (err, result) {
-            if (err) {
-                throw (err);
-            } else if (result.affectedRows === 1) {
-                res.send(true);
-            } else {
-                throw (new ReferenceError('No such notification or notification already unread'));
-            }
-        });
-    } catch (err) {
-        res.redirect(utils.errorRedirect('/notifications/markUnread', 'An unexpected error occurred marking your notification as unread.'));
-    }
-});
-
-
-//delete all notifications for a user.
-app.delete('/notifications/delete', utils.checkAuthenticated, (req, res, next) => {
-    try {
-        let sql = 'DELETE FROM Notifications WHERE UserID=?';
-        mysql.pool.query(sql, [req.user.UserKey], function (err, result) {
-            if (err) {
-                throw (err);
-            } else if (result.affectedRows >= 1) {
-                res.send(true);
-            } else {
-                res.send(true); //don't worry about throwing an error. could be that they deleted on one tab and tried to do it again on another
-            }
-        });
-    } catch (err) {
-        res.redirect(utils.errorRedirect('/notifications/delete', 'An unexpected error occurred deleting your notifications.'));
-    }
-});
-
-//delete a single notification for a user.
-app.delete('/notifications/delete/:id', utils.checkAuthenticated, (req, res, next) => {
-    try {
-        let sql = 'DELETE FROM Notifications WHERE UserID=? AND NotificationKey=?';
-        mysql.pool.query(sql, [req.user.UserKey, req.params.id], function (err, result) {
-            if (err) {
-                throw (err);
-            } else if (result.affectedRows === 1) {
-                res.send(true);
-            } else {
-                res.send(true); //don't worry about throwing an error. could be that they deleted on one tab and tried to do it again on another
-            }
-        });
-    } catch (err) {
-        res.redirect(utils.errorRedirect('/notifications/delete', 'An unexpected error occurred deleting a notification.'));
-    }
 });
 
 // start app
